@@ -18,8 +18,9 @@
     "deploy:check": "cross-env NODE_ENV=production node deployment-check.js",
     "deploy:check:win": "set NODE_ENV=production && node deployment-check.js",
     "security:check": "npm audit && npm run lint",
-    "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
-    "lint:fix": "eslint . --ext .js,.jsx,.ts,.tsx --fix",
+  "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
+  "lint:fix": "eslint . --ext .js,.jsx,.ts,.tsx --fix",
+  "lint:report": "eslint . -f stylish -o tmp-eslint-output.txt",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
     "test": "jest --env=jsdom --config jest.config.js && cypress run",
@@ -46,17 +47,21 @@
     "Edge >= 79"
   ],
   "devDependencies": {
-    "@babel/eslint-parser": "^7.28.0",
-    "@babel/preset-env": "^7.28.3",
-    "@babel/preset-react": "^7.27.1",
+  "@babel/eslint-parser": "^7.28.0",
+  "@babel/preset-env": "^7.28.3",
+  "@babel/preset-react": "^7.27.1",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
     "@vitejs/plugin-react": "^4.7.0",
     "autoprefixer": "^10.4.21",
     "cross-env": "^7.0.3",
-    "eslint": "^8.56.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-react": "^7.33.0",
+  "eslint": "^9.0.0",
+  "@eslint/js": "^1.0.0",
+  "eslint-config-prettier": "^9.1.0",
+  "eslint-plugin-react": "^7.33.0",
+  "eslint-plugin-react-hooks": "^4.6.0",
+  "eslint-plugin-jsx-a11y": "^6.7.1",
+  "eslint-plugin-import": "^2.29.1",
     "husky": "^9.1.7",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^30.0.3",
